+++ src/forms/upload/multi-upload/multi-upload.component.js
@@ -14,7 +14,7 @@ const Component = {
     icon: '@', // illustration in icon shown in upload box
 
     buttonText: '@', // Button text shown in default state
-    secondaryButtonText: '@', // Button text shown in processing state
+    addMoreButtonText: '@', // Button text shown when at least one file is uploading
 
     droppingText: '@', // Text shown when hovering with a file
     processingText: '@', // Text shown while processing/uploading
@@ -28,6 +28,7 @@ const Component = {
 
     onStart: '&',
     onFinish: '&',
+    onFailure: '&',
 
     accept: '@',
     httpOptions: '<',